requirements.txt
Flask>=0.10.1
pam>=0.1.4
Pygments>=2.0.2
python-magic>=0.4.6
